pyproject.toml
[project]
name = "kn-api-server"
version = "1.0.0"
description = "A simple FastAPI service with an echo endpoint"
authors = [
    {name = "Developer", email = "developer@example.com"},
]
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "google-api-python-client>=2.181.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "authlib>=1.6.3",
    "requests>=2.31.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.32",
    "tiktoken>=0.7.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
Homepage = "https://github.com/kos2nov/divizapi"
Repository = "https://github.com/kos2nov/divizapi.git"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["diviz"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
echo-start = "diviz.main:main"



[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--disable-warnings"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests"
]
